重构提取redis抽象层到weixin-java-common模块
