+++ weixin-java-common/pom.xml
@@ -122,10 +122,17 @@
       <groupId>redis.clients</groupId>
       <artifactId>jedis</artifactId>
     </dependency>
-
     <dependency>
       <groupId>com.github.jedis-lock</groupId>
       <artifactId>jedis-lock</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.redisson</groupId>
+      <artifactId>redisson</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.data</groupId>
+      <artifactId>spring-data-redis</artifactId>
     </dependency>
   </dependencies>
