+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/redis/BaseWxRedisOps.java
@@ -1,14 +1,14 @@
-package me.chanjar.weixin.mp.config.redis;
+package me.chanjar.weixin.common.redis;
 
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Lock;
 
 /**
- * 微信公众号redis操作基本类
+ * 微信redis操作基本类
  * <p>
  * 非内置实现redis相关操作, 请实现该类
  */
-public class BaseWxMpRedisOps implements WxMpRedisOps {
+public abstract class BaseWxRedisOps implements WxRedisOps {
 
   @Override
   public String getValue(String key) {
